Refactored Tower script

+++ GFC Submission/Assets/Scripts/Tower Defense/Enemies/Enemy.cs
@@ -33,7 +33,7 @@ public class Enemy : MonoBehaviour
         return trackSectionIndex + trackSectionPos;
     }
 
-    public void TakeDamage(int dmg)
+    public void TakeDamage(float dmg)
     {
         health -= dmg;
         if (health <= 0)
